LYMON-891 refactor: extract RoleAssignmentValidator and make supplier items pagination optional
Extract duplicated role assignment validation logic from InviteStaffHandler
and UpdateStaffHandler into a shared RoleAssignmentValidator service.
Make page and limit query params optional in GET /suppliers/:supplierId/items.
