+++ src/application/application.module.ts
@@ -42,6 +42,7 @@ import { CartApplicationModule } from '@/application/cart/cart-application.modul
 import { RefundApplicationModule } from '@/application/refund/refund-application.module';
 import { ProcessWompiWebhookHandler } from '@/application/payment/commands/process-wompi-webhook/process-wompi-webhook.handler';
 import { GetPaymentSessionStatusHandler } from '@/application/payment/queries/get-payment-session-status/get-payment-session-status.handler';
+import { RoleAssignmentValidator } from '@/application/user/services/role-assignment-validator.service';
 
 const CommandHandlers = [
   RegisterTenantHandler,
@@ -92,7 +93,7 @@ const QueryHandlers = [GetShiftsHandler];
     CartApplicationModule,
     RefundApplicationModule,
   ],
-  providers: [...CommandHandlers, ...QueryHandlers],
+  providers: [...CommandHandlers, ...QueryHandlers, RoleAssignmentValidator],
   exports: [
     ...CommandHandlers,
     ...QueryHandlers,
