+++ src/application/user/commands/invite-staff/invite-staff.handler.ts
@@ -17,23 +17,10 @@ import {
   type TenantRepository,
 } from '@/domain/tenant/repositories/tenant.repository';
 import {
-  PROPERTY_REPOSITORY,
-  type PropertyRepository,
-} from '@/domain/property/repositories/property.repository';
-import {
-  UNIT_REPOSITORY,
-  type UnitRepository,
-} from '@/domain/unit/repositories/unit.repository';
-import {
-  ROLE_REPOSITORY,
-  type RoleRepository,
-} from '@/domain/role/repositories/role.repository';
-import { RoleId } from '@/domain/role/entities/role.entity';
-import {
   type IPasswordHasher,
   PASSWORD_HASHER,
 } from '@/application/auth/services/password-hasher.service';
-import { RoleAssignment, User } from '@/domain/user/entities/user.entity';
+import { User } from '@/domain/user/entities/user.entity';
 import { EventEmitter2 } from '@nestjs/event-emitter';
 import {
   AuditLoggedEvent,
@@ -43,6 +30,7 @@ import {
   AuditAction,
   AuditEntityType,
 } from '@/domain/audit/value-objects/audit-action.vo';
+import { RoleAssignmentValidator } from '@/application/user/services/role-assignment-validator.service';
 
 @CommandHandler(InviteStaffCommand)
 export class InviteStaffHandler implements ICommandHandler<InviteStaffCommand> {
@@ -51,14 +39,9 @@ export class InviteStaffHandler implements ICommandHandler<InviteStaffCommand> {
     private readonly userRepository: UserRepository,
     @Inject(TENANT_REPOSITORY)
     private readonly tenantRepository: TenantRepository,
-    @Inject(PROPERTY_REPOSITORY)
-    private readonly propertyRepository: PropertyRepository,
-    @Inject(UNIT_REPOSITORY)
-    private readonly unitRepository: UnitRepository,
-    @Inject(ROLE_REPOSITORY)
-    private readonly roleRepository: RoleRepository,
     @Inject(PASSWORD_HASHER)
     private readonly passwordHasher: IPasswordHasher,
+    private readonly roleAssignmentValidator: RoleAssignmentValidator,
     private readonly eventEmitter: EventEmitter2,
   ) {}
 
@@ -83,7 +66,7 @@ export class InviteStaffHandler implements ICommandHandler<InviteStaffCommand> {
     }
 
     await this.validatePlanLimits(tenantId, tenant.getPlan().getStaffLimit());
-    await this.validateRoleAssignments(
+    await this.roleAssignmentValidator.validate(
       command.roleAssignments,
       tenantId.toString(),
     );
@@ -115,78 +98,6 @@ export class InviteStaffHandler implements ICommandHandler<InviteStaffCommand> {
         { invitedEmail: command.email },
       ),
     );
-  }
-
-  private async validateRoleAssignments(
-    assignments: RoleAssignment[],
-    tenantId: string,
-  ): Promise<void> {
-    let validPropertyIds: Set<string> | null = null;
-    let validUnitIds: Set<string> | null = null;
-
-    for (const assignment of assignments) {
-      const roleId = RoleId.createFromString(assignment.roleId);
-      const role = await this.roleRepository.findById(roleId);
-      if (!role) {
-        throw new BadRequestException(
-          `Role '${assignment.roleId}' does not exist`,
-        );
-      }
-
-      if (assignment.scope.type === 'PROPERTY') {
-        validPropertyIds = await this.validatePropertyScope(
-          assignment.scope.resourceIds,
-          tenantId,
-          validPropertyIds,
-        );
-      }
-
-      if (assignment.scope.type === 'UNIT') {
-        validUnitIds = await this.validateUnitScope(
-          assignment.scope.resourceIds,
-          tenantId,
-          validUnitIds,
-        );
-      }
-    }
-  }
-
-  private async validatePropertyScope(
-    resourceIds: string[],
-    tenantId: string,
-    cache: Set<string> | null,
-  ): Promise<Set<string>> {
-    if (!cache) {
-      const tid = TenantId.createFromString(tenantId);
-      const properties = await this.propertyRepository.findByTenantId(tid);
-      cache = new Set(properties.map((p) => p.getId()!.toString()));
-    }
-    const invalid = resourceIds.filter((id) => !cache.has(id));
-    if (invalid.length > 0) {
-      throw new BadRequestException(
-        `Property IDs not found in this tenant: ${invalid.join(', ')}`,
-      );
-    }
-    return cache;
-  }
-
-  private async validateUnitScope(
-    resourceIds: string[],
-    tenantId: string,
-    cache: Set<string> | null,
-  ): Promise<Set<string>> {
-    if (!cache) {
-      const tid = TenantId.createFromString(tenantId);
-      const units = await this.unitRepository.findByTenantId(tid);
-      cache = new Set(units.map((u) => u.getId()!.toString()));
-    }
-    const invalid = resourceIds.filter((id) => !cache.has(id));
-    if (invalid.length > 0) {
-      throw new BadRequestException(
-        `Unit IDs not found in this tenant: ${invalid.join(', ')}`,
-      );
-    }
-    return cache;
   }
 
   private async validatePlanLimits(tenantId: TenantId, staffLimit: number) {
