+++ src/application/user/commands/update-staff.handler.ts
@@ -8,19 +8,6 @@ import {
 import { RoleAssignment } from '@/domain/user/entities/user.entity';
 import { UserId } from '@/domain/user/entities/user.entity';
 import { TenantId } from '@/domain/tenant/value-objects/tenant-id.vo';
-import {
-  PROPERTY_REPOSITORY,
-  type PropertyRepository,
-} from '@/domain/property/repositories/property.repository';
-import {
-  UNIT_REPOSITORY,
-  type UnitRepository,
-} from '@/domain/unit/repositories/unit.repository';
-import {
-  ROLE_REPOSITORY,
-  type RoleRepository,
-} from '@/domain/role/repositories/role.repository';
-import { RoleId } from '@/domain/role/entities/role.entity';
 import { EventEmitter2 } from '@nestjs/event-emitter';
 import {
   AuditLoggedEvent,
@@ -30,18 +17,14 @@ import {
   AuditAction,
   AuditEntityType,
 } from '@/domain/audit/value-objects/audit-action.vo';
+import { RoleAssignmentValidator } from '@/application/user/services/role-assignment-validator.service';
 
 @CommandHandler(UpdateStaffCommand)
 export class UpdateStaffHandler implements ICommandHandler<UpdateStaffCommand> {
   constructor(
     @Inject(USER_REPOSITORY)
     private readonly userRepository: UserRepository,
-    @Inject(PROPERTY_REPOSITORY)
-    private readonly propertyRepository: PropertyRepository,
-    @Inject(UNIT_REPOSITORY)
-    private readonly unitRepository: UnitRepository,
-    @Inject(ROLE_REPOSITORY)
-    private readonly roleRepository: RoleRepository,
+    private readonly roleAssignmentValidator: RoleAssignmentValidator,
     private readonly eventEmitter: EventEmitter2,
   ) {}
 
@@ -73,7 +56,7 @@ export class UpdateStaffHandler implements ICommandHandler<UpdateStaffCommand> {
 
     if (command.roleAssignments && command.roleAssignments.length > 0) {
       // Full replacement after validation
-      await this.validateRoleAssignments(
+      await this.roleAssignmentValidator.validate(
         command.roleAssignments,
         user.getTenantId().toString(),
       );
@@ -81,7 +64,7 @@ export class UpdateStaffHandler implements ICommandHandler<UpdateStaffCommand> {
     } else {
       // apply add/remove
       if (command.addPermissions && command.addPermissions.length > 0) {
-        await this.validateRoleAssignments(
+        await this.roleAssignmentValidator.validate(
           command.addPermissions,
           user.getTenantId().toString(),
         );
@@ -150,77 +133,5 @@ export class UpdateStaffHandler implements ICommandHandler<UpdateStaffCommand> {
       a.roleId === b.roleId &&
       JSON.stringify(a.scope) === JSON.stringify(b.scope)
     );
-  }
-
-  private async validateRoleAssignments(
-    assignments: RoleAssignment[],
-    tenantId: string,
-  ) {
-    let validPropertyIds: Set<string> | null = null;
-    let validUnitIds: Set<string> | null = null;
-
-    for (const assignment of assignments) {
-      const roleId = RoleId.createFromString(assignment.roleId);
-      const role = await this.roleRepository.findById(roleId);
-      if (!role) {
-        throw new BadRequestException(
-          `Role '${assignment.roleId}' does not exist`,
-        );
-      }
-
-      if (assignment.scope.type === 'PROPERTY') {
-        validPropertyIds = await this.validatePropertyScope(
-          assignment.scope.resourceIds,
-          tenantId,
-          validPropertyIds,
-        );
-      }
-
-      if (assignment.scope.type === 'UNIT') {
-        validUnitIds = await this.validateUnitScope(
-          assignment.scope.resourceIds,
-          tenantId,
-          validUnitIds,
-        );
-      }
-    }
-  }
-
-  private async validatePropertyScope(
-    resourceIds: string[],
-    tenantId: string,
-    cache: Set<string> | null,
-  ): Promise<Set<string>> {
-    if (!cache) {
-      const tid = TenantId.createFromString(tenantId);
-      const properties = await this.propertyRepository.findByTenantId(tid);
-      cache = new Set(properties.map((p) => p.getId()!.toString()));
-    }
-    const invalid = resourceIds.filter((id) => !cache.has(id));
-    if (invalid.length > 0) {
-      throw new BadRequestException(
-        `Property IDs not found in this tenant: ${invalid.join(', ')}`,
-      );
-    }
-    return cache;
-  }
-
-  private async validateUnitScope(
-    resourceIds: string[],
-    tenantId: string,
-    cache: Set<string> | null,
-  ): Promise<Set<string>> {
-    if (!cache) {
-      const tid = TenantId.createFromString(tenantId);
-      const units = await this.unitRepository.findByTenantId(tid);
-      cache = new Set(units.map((u) => u.getId()!.toString()));
-    }
-    const invalid = resourceIds.filter((id) => !cache.has(id));
-    if (invalid.length > 0) {
-      throw new BadRequestException(
-        `Unit IDs not found in this tenant: ${invalid.join(', ')}`,
-      );
-    }
-    return cache;
   }
 }
